refactored some classes #2
